refactor(plan_SERVERSPLIT slice 9): the cdp-functions bodies leave server.py
The thirteen `cdp-functions` tool bodies move VERBATIM out of
embedded/server.py into embedded/tool_sections/cdp_functions.py. Pure move: the
only edits are the dropped registration decorator (contract rule 2) and the
rewrite of the singleton/helper reads to `rt.<name>`, resolved at CALL time
against the one patchable home (contract rule 3).

This is the one section a runtime GATE switches off. Both gate sites STAY in
server.py — the module-scope `settings.xpool_safe_mode` branch and the __main__
block's --xpool-safe / --disable-cdp-functions — and both still run AFTER the
binding loop, which is what plan_SERVERSPLIT R6 requires: apply_disabled_sections
works by mcp.remove_tool, so it can only remove tools that are already
registered. A loop placed after a gate would leave the thirteen registered into
a gate that had already closed.

Verified against a REAL backend subprocess over HTTP rather than in process
(spawned on a throwaway port with HOME/USERPROFILE redirected at a temp dir, so
the real ~/.stealth-mcp is untouched), pre-move and post-move alike:

  control                 SERVED 94
  --xpool-safe            SERVED 81   (13 removed)
  --disable-cdp-functions SERVED 81   (13 removed)
  XPOOL_SAFE_MODE=1       SERVED 81   (13 removed, the module-scope branch)

Two guards moved with the bodies and are updated in this commit rather than
allowed to go stale:

  * release_evidence.HOST_EXEC_SITES named embedded/server.py as a host-Python
    execution site because of create_python_binding's `exec`. The SITE is
    unchanged — same tool, same caller text, same host privileges — but its file
    is now the section module, which is exactly the reviewed-decision-not-a-
    discovery the inventory exists for. RELEASE_CONTRACT.md does not name the
    file, so gen_release_contract --check stays clean.
  * §5.5's MIGRATION_ALIASES floor was 15. Slices 7, 8 and 9 each pruned the
    last server.py consumer of a tool_runtime alias (file_based_element_cloner,
    cdp_element_cloner, cdp_function_executor), so the derived set is 14. The
    floor RATCHETS DOWN to the measured actual — the mirror of server.py's LOC
    cap and the opposite of source_scan's floor — so the pin still cannot vanish
    silently while a deliberate prune is not read as one.

server.py 1748 -> 1371 LOC (376 body lines plus the one-line
`cdp_function_executor` alias import); the GRANDFATHER cap ratchets DOWN to the
measured post-ruff-format actual, and tests/source_scan.py's floor ratchets UP
10 -> 11.

Verified: wire surface IDENTICAL to tests/goldens/tool_surface.json;
SECTION_TOOLS 94 with per-app [94, 94, 94]; `--list-sections` subprocess
Total: 94 tools with cdp-functions at 13; unit lane 2329 passed / 1 skipped;
real-Chrome test_e2e_functions_hooks.py 7 passed and test_e2e_transport.py
1 passed; budgets, ruff format+check, suppression owners, vulture, ty
(88 warnings, 0 errors) and gen_release_contract --check all clean.
